LPS-89617 Allow usage of widgets in the content page templates and display pages
