+++ modules/apps/layout/layout-content-page-editor-web/src/main/java/com/liferay/layout/content/page/editor/web/internal/display/context/ContentPageEditorLayoutPageTemplateDisplayContext.java
@@ -92,7 +92,7 @@ public class ContentPageEditorLayoutPageTemplateDisplayContext
 		}
 
 		soyContext.put(
-			"sidebarPanels", getSidebarPanelSoyContexts(_showMapping, false));
+			"sidebarPanels", getSidebarPanelSoyContexts(_showMapping, true));
 
 		LayoutPageTemplateEntry layoutPageTemplateEntry =
 			_getLayoutPageTemplateEntry();
