+++ modules/apps/layout/layout-content-page-editor-web/src/main/java/com/liferay/layout/content/page/editor/web/internal/display/context/ContentPageLayoutEditorDisplayContext.java
@@ -16,45 +16,15 @@ package com.liferay.layout.content.page.editor.web.internal.display.context;
 
 import com.liferay.fragment.model.FragmentEntry;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.Portlet;
-import com.liferay.portal.kernel.model.PortletApp;
-import com.liferay.portal.kernel.model.PortletCategory;
-import com.liferay.portal.kernel.portlet.PortletConfigFactoryUtil;
 import com.liferay.portal.kernel.portlet.PortletIdCodec;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.service.PortletLocalServiceUtil;
-import com.liferay.portal.kernel.service.PortletPreferencesLocalServiceUtil;
-import com.liferay.portal.kernel.service.permission.PortletPermissionUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.PortletKeys;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.kernel.util.comparator.PortletCategoryComparator;
-import com.liferay.portal.kernel.util.comparator.PortletTitleComparator;
 import com.liferay.portal.template.soy.util.SoyContext;
 import com.liferay.portal.template.soy.util.SoyContextFactoryUtil;
-import com.liferay.portal.util.PortletCategoryUtil;
-import com.liferay.portal.util.WebAppPool;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.ResourceBundle;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import javax.portlet.PortletConfig;
 import javax.portlet.RenderResponse;
 
-import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
@@ -85,9 +55,9 @@ public class ContentPageLayoutEditorDisplayContext
 		soyContext.put(
 			"addPortletURL",
 			getFragmentEntryActionURL("/content_layout/add_portlet"));
+
 		soyContext.put(
 			"sidebarPanels", getSidebarPanelSoyContexts(false, true));
-		soyContext.put("widgets", _getWidgetsSoyContexts());
 
 		_editorSoyContext = soyContext;
 
@@ -107,7 +77,7 @@ public class ContentPageLayoutEditorDisplayContext
 		soyContext.put("fragmentEntryId", 0);
 		soyContext.put("name", StringPool.BLANK);
 
-		String portletId = _getPortletId(content);
+		String portletId = getPortletId(content);
 
 		if (Validator.isNull(portletId)) {
 			return soyContext;
@@ -121,43 +91,7 @@ public class ContentPageLayoutEditorDisplayContext
 		return soyContext;
 	}
 
-	private String _getPortletCategoryTitle(PortletCategory portletCategory) {
-		String title = LanguageUtil.get(request, portletCategory.getName());
-
-		if (Validator.isNotNull(title)) {
-			return title;
-		}
-
-		for (String portletId :
-				PortletCategoryUtil.getFirstChildPortletIds(portletCategory)) {
-
-			Portlet portlet = PortletLocalServiceUtil.getPortletById(
-				themeDisplay.getCompanyId(), portletId);
-
-			if (portlet == null) {
-				continue;
-			}
-
-			PortletApp portletApp = portlet.getPortletApp();
-
-			if (!portletApp.isWARFile()) {
-				continue;
-			}
-
-			PortletConfig portletConfig = PortletConfigFactoryUtil.create(
-				portlet, request.getServletContext());
-
-			ResourceBundle portletResourceBundle =
-				portletConfig.getResourceBundle(themeDisplay.getLocale());
-
-			return ResourceBundleUtil.getString(
-				portletResourceBundle, portletCategory.getName());
-		}
-
-		return StringPool.BLANK;
-	}
-
-	private String _getPortletId(String content) {
+	protected String getPortletId(String content) {
 		Document document = Jsoup.parse(content);
 
 		Elements elements = document.getElementsByAttributeValueStarting(
@@ -173,131 +107,6 @@ public class ContentPageLayoutEditorDisplayContext
 
 		return PortletIdCodec.decodePortletName(id.substring(8));
 	}
-
-	private List<SoyContext> _getPortletsContexts(
-		PortletCategory portletCategory) {
-
-		Set<String> portletIds = portletCategory.getPortletIds();
-
-		Stream<String> stream = portletIds.stream();
-
-		HttpSession session = request.getSession();
-
-		ServletContext servletContext = session.getServletContext();
-
-		return stream.map(
-			portletId -> PortletLocalServiceUtil.getPortletById(
-				themeDisplay.getCompanyId(), portletId)
-		).filter(
-			portlet -> {
-				if (portlet == null) {
-					return false;
-				}
-
-				try {
-					return PortletPermissionUtil.contains(
-						themeDisplay.getPermissionChecker(),
-						themeDisplay.getLayout(), portlet,
-						ActionKeys.ADD_TO_PAGE);
-				}
-				catch (PortalException pe) {
-					_log.error(
-						"Unable to check portlet permissions for " +
-							portlet.getPortletId(),
-						pe);
-
-					return false;
-				}
-			}
-		).sorted(
-			new PortletTitleComparator(servletContext, themeDisplay.getLocale())
-		).map(
-			portlet -> {
-				SoyContext portletSoyContext =
-					SoyContextFactoryUtil.createSoyContext();
-
-				portletSoyContext.put("instanceable", portlet.isInstanceable());
-				portletSoyContext.put("portletId", portlet.getPortletId());
-				portletSoyContext.put(
-					"title",
-					PortalUtil.getPortletTitle(
-						portlet, servletContext, themeDisplay.getLocale()));
-				portletSoyContext.put(
-					"used", _isUsed(portlet, themeDisplay.getPlid()));
-
-				return portletSoyContext;
-			}
-		).collect(
-			Collectors.toList()
-		);
-	}
-
-	private List<SoyContext> _getWidgetCategoriesContexts(
-		PortletCategory portletCategory) {
-
-		Collection<PortletCategory> portletCategories =
-			portletCategory.getCategories();
-
-		Stream<PortletCategory> stream = portletCategories.stream();
-
-		return stream.sorted(
-			new PortletCategoryComparator(themeDisplay.getLocale())
-		).filter(
-			category -> !category.isHidden()
-		).map(
-			category -> {
-				SoyContext categoryContext =
-					SoyContextFactoryUtil.createSoyContext();
-
-				categoryContext.put(
-					"categories", _getWidgetCategoriesContexts(category));
-				categoryContext.put(
-					"path",
-					StringUtil.replace(
-						category.getPath(), new String[] {"/", "."},
-						new String[] {"-", "-"}));
-				categoryContext.put("portlets", _getPortletsContexts(category));
-				categoryContext.put(
-					"title", _getPortletCategoryTitle(category));
-
-				return categoryContext;
-			}
-		).collect(
-			Collectors.toList()
-		);
-	}
-
-	private List<SoyContext> _getWidgetsSoyContexts() throws Exception {
-		PortletCategory portletCategory = (PortletCategory)WebAppPool.get(
-			themeDisplay.getCompanyId(), WebKeys.PORTLET_CATEGORY);
-
-		portletCategory = PortletCategoryUtil.getRelevantPortletCategory(
-			themeDisplay.getPermissionChecker(), themeDisplay.getCompanyId(),
-			themeDisplay.getLayout(), portletCategory,
-			themeDisplay.getLayoutTypePortlet());
-
-		return _getWidgetCategoriesContexts(portletCategory);
-	}
-
-	private boolean _isUsed(Portlet portlet, long plid) {
-		if (portlet.isInstanceable()) {
-			return false;
-		}
-
-		long count =
-			PortletPreferencesLocalServiceUtil.getPortletPreferencesCount(
-				PortletKeys.PREFS_OWNER_TYPE_LAYOUT, plid,
-				portlet.getPortletId());
-
-		if (count > 0) {
-			return true;
-		}
-
-		return false;
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		ContentPageLayoutEditorDisplayContext.class);
 
 	private SoyContext _editorSoyContext;
